move home component to pages folder
